Creating controller folder
